backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg==3.2.4
celery==5.3.4
redis==5.0.1
beautifulsoup4==4.12.2
psycopg-binary==3.2.10
requests==2.31.0
newspaper3k==0.2.8
openai==1.3.7
sentence-transformers==2.2.2
pinecone-client==2.2.4
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4